Integration playwright tests reorganisation and error state update
